requirements.txt
appdirs==1.4.3
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.11
djangorestframework==3.11.0
filelock==3.0.12
Flask==1.0.2
gunicorn==20.0.4
idna==2.6
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.16.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.6.2
passlib==1.7.2
pbr==5.4.4
peewee==3.3.4
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2-binary==2.8.5
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2020.1
requests==2.18.4
six==1.14.0
stevedore==1.32.0
traitlets==4.3.3
urllib3==1.22
virtualenv==20.0.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
Werkzeug==0.14.1
